Modify what the buttons do and how they do it
Deleting the grid squares was previously handled by the clear function,
which responded to a click on the clear button. However, this required
the user to always clear the grid prior to setting a new size.

I changed the functionality of the size button to clear the grid
(delete all the grid squares) before creating a new grid - whenever a
new size is entered.

The clear function, now, just whites out a grid (sets the
color of all grid squares to white).

I also refactored the code for the two color options so that event
listeners aren't unnecessarily created each time the color is changed.